+++ src/constants/defaultOptions.ts
@@ -1,0 +1,8 @@
+import { Options } from "../types/options"
+import { headersToCheck } from "./headerlist"
+
+export const defaultOptions: Options = {
+  headersOnly: false,
+  checkHeaders: headersToCheck,
+  injectServer: (app) => app.server,
+}
