+++ src/index.ts
@@ -1,95 +1,9 @@
-import { Server } from "bun"
-import { Elysia } from "elysia"
+export { getIP } from "./services/getip"
+export { plugin as ip } from "./services/plugin"
 
-export type IPHeaders =
-  | "x-real-ip"
-  | "x-client-ip"
-  | "cf-connecting-ip"
-  | "fastly-client-ip"
-  | "x-cluster-client-ip"
-  | "x-forwarded"
-  | "forwarded-for"
-  | "forwarded"
-  | "x-forwarded"
-  | "appengine-user-ip"
-  | "true-client-ip"
-  | "cf-pseudo-ipv4"
-  | (string & {})
-export const headersToCheck: IPHeaders[] = [
-  "x-real-ip", // Nginx proxy/FastCGI
-  "x-client-ip", // Apache https://httpd.apache.org/docs/2.4/mod/mod_remoteip.html#page-header
-  "cf-connecting-ip", // Cloudflare
-  "fastly-client-ip", // Fastly
-  "x-cluster-client-ip", // GCP
-  "x-forwarded", // General Forwarded
-  "forwarded-for", // RFC 7239
-  "forwarded", // RFC 7239
-  "x-forwarded", // RFC 7239
-  "appengine-user-ip", // GCP
-  "true-client-ip", // Akamai and Cloudflare
-  "cf-pseudo-ipv4", // Cloudflare
-]
+export { defaultOptions } from "./constants/defaultOptions"
+export { headersToCheck } from "./constants/headerlist"
 
-export const getIP = (
-  headers: Headers,
-  checkHeaders: IPHeaders[] = headersToCheck,
-) => {
-  if (typeof checkHeaders === "string" && headers.get(checkHeaders))
-    return headers.get(checkHeaders)
-  // X-Forwarded-For is the de-facto standard header
-  if (headers.get("x-forwarded-for"))
-    return headers.get("x-forwarded-for")?.split(",")[0]
-  let clientIP: string | undefined | null = null
-  if (!checkHeaders) return null
-  for (const header of checkHeaders) {
-    clientIP = headers.get(header)
-    if (clientIP) break
-  }
-  return clientIP
-}
-
-export const ip =
-  (
-    config: {
-      /**
-       * Customize headers to check for IP address
-       * @default ['x-real-ip', 'x-client-ip', 'cf-connecting-ip', 'fastly-client-ip', 'x-cluster-client-ip', 'x-forwarded', 'forwarded-for', 'forwarded', 'x-forwarded', 'appengine-user-ip', 'true-client-ip', 'cf-pseudo-ipv4']
-       */
-      checkHeaders?: IPHeaders[]
-      /**
-       * Only check headers regardless of the runtime environment
-       * @default false
-       */
-      headersOnly?: boolean
-      /**
-       *
-       */
-      injectServer: (app: Elysia) => Server | null
-    } = {
-      headersOnly: false,
-      injectServer: (app) => app.server,
-    },
-  ) =>
-  (app: Elysia) =>
-    new Elysia({
-      name: "elysia-ip",
-    }).derive({ as: "global" }, ({ request }): { ip: string } => {
-      if (!config.headersOnly && globalThis.Bun) {
-        const socketAddress = config.injectServer(app)?.requestIP(request)
-        if (socketAddress) return { ip: socketAddress.address }
-        else {
-          console.log(
-            `Elysia server is not initialized. Make sure to call Elyisa.listen()`,
-          )
-          console.log(`use injectServer to inject Server instance`)
-        }
-      }
-      return {
-        ip:
-          getIP(request.headers, config.checkHeaders) ||
-          (function () {
-            console.log("Failed to get ip from header!")
-            return ""
-          })(),
-      }
-    })
+export type { IPHeaders } from "./types/header"
+export type { Options } from "./types/options"
+export type { injectServer } from "./types/injectserver"
