+++ src/types/injectserver.ts
@@ -1,0 +1,4 @@
+import type { Server } from "bun"
+import type Elysia from "elysia"
+
+export type injectServer = (app: Elysia) => Server | null
